refactor: npc ships all inherit from AbstractShip
So far, all npc ships have a navigation system with avoidance. As a side
effect, the mothership now also use navigation, and aims at a Target
node in World.

The radius of avoidance is overwritten in code by the radius a
AvoidanceCircle: an abstract collisionShape that should be defined as a
circle on each inheriting Ship